updated Route4MeDB project

+++ Route4MeDB/DatabaseMirgation/Program.cs
@@ -18,6 +18,9 @@ using Microsoft.AspNetCore.Builder.Internal;
 using Microsoft.EntityFrameworkCore;
 using System.Threading.Tasks;
 using Route4MeDB.UnitTests.Builders;
+using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Storage;
+using Route4MeDbLibrary;
 
 namespace RouteMeDB.DatabaseMirgation
 {
@@ -25,31 +28,62 @@ namespace RouteMeDB.DatabaseMirgation
     {
         static void Main(string[] args)
         {
-            
-
+            var dbProvider = DatabaseProviders.PostgreSql;
             var curPath = Directory.GetCurrentDirectory();
             var configBuilder = new ConfigurationBuilder()
                .SetBasePath(curPath)
                .AddJsonFile("appsettings.json", optional: true);
             var config = configBuilder.Build();
-            
-            Route4MeDbManager.DatabaseProvider = DatabaseProviders.LocalDb;
+            //var basedir = Utils.GetRoute4MeDbBaseDir(curPath);
+            // Specify here a database provider
+
+            Route4MeDbManager.DatabaseProvider = dbProvider;
 
             var r4mdbManager = new Route4MeDbManager(config);
            
             _route4meDbContext = r4mdbManager.Route4MeContext;
 
-            var optBuilder = new DbContextOptionsBuilder<Route4MeDbContext>()
-                .UseSqlServer(config.GetConnectionString("DefaultConnection"));
+            //Specify here the parameters for appropriate DB enigne
+            var optBuilder= new DbContextOptionsBuilder<Route4MeDbContext>();
+
+            switch (dbProvider)
+            {
+                case DatabaseProviders.LocalDb:
+                    optBuilder.UseSqlServer(config.GetConnectionString("DefaultConnection"));
+                    break;
+                case DatabaseProviders.Sqlexpress:
+                    optBuilder.UseSqlServer(config.GetConnectionString("SqlExpressConnection"));
+                    break;
+                case DatabaseProviders.MySql:
+                    optBuilder.UseMySql(config.GetConnectionString("MySqlConnection"));
+                    break;
+                case DatabaseProviders.SQLite:
+                    optBuilder.UseSqlite(config.GetConnectionString("SQLiteConnection"));
+                    break;
+                case DatabaseProviders.PostgreSql:
+                    optBuilder.UseNpgsql(config.GetConnectionString("PostgreSqlConnection"));
+                    break;
+            }
+
+            //var optBuilder = new DbContextOptionsBuilder<Route4MeDbContext>()
+            //    .UseMySQL(config.GetConnectionString("MySqlConnection"));
 
             var services = new ServiceCollection()
             .AddLogging().AddSingleton<DbContextOptions>(optBuilder.Options)
             .AddOptions()
-            .AddDbContext<Route4MeDbContext>().AddEntityFrameworkSqlServer()
+            .AddDbContext<Route4MeDbContext>()
             .BuildServiceProvider();
 
-            _route4meDbContext.Database.EnsureDeleted();
+            if ((_route4meDbContext.Database.GetService<IDatabaseCreator>() as RelationalDatabaseCreator).Exists())
+                _route4meDbContext.Database.EnsureDeleted();
+
             _route4meDbContext.Database.EnsureCreated();
+
+            if (dbProvider == DatabaseProviders.MySql)
+            {
+                string command = @"CREATE TABLE `__EFMigrationsHistory` ( `MigrationId` nvarchar(150) NOT NULL, `ProductVersion` nvarchar(32) NOT NULL, PRIMARY KEY (`MigrationId`) );";
+                _route4meDbContext.Database.ExecuteSqlCommandAsync(command).Wait();
+            }
 
             IApplicationBuilder app = new ApplicationBuilder(services);
 
@@ -70,6 +104,18 @@ namespace RouteMeDB.DatabaseMirgation
         private static AddressBookContactBuilder contactBuilder { get; } = new AddressBookContactBuilder();
         static Route4MeDbContext _route4meDbContext;
         
+        private string GetRoute4MeDbBaseDir(string curPath)
+        {
+            var dirinfo = new DirectoryInfo(curPath);
+
+            while (dirinfo.Name != "Route4MeDB")
+            {
+                curPath = dirinfo.Parent.FullName;
+                dirinfo = new DirectoryInfo(curPath);
+            }
+
+            return curPath;
+        }
 
         public async Task<AddressBookContact> CreateContact()
         {
